Use alternate approach for defining flag CPs
Because the LD `allFlags` function calls through the `variation` which
ultimately sends an event request that updates the "last reqauested"
timestamp of each flag, we needed to find an alternative approach to
defining the CP's for each flag.

Instead of cycling through the list of all flags retrieved from LD via
the method above, and creating a CP for each, this commit will us the
`unknownProperty` to dynamically record the flag being requested and
create a CP for it.

If something then happens that should invalidate the CP's, such as
identifying a user, the code can cycle through the list of seen flags
and `notifyPropertyChange` for each one.