Pass flag key as default value
This is used so that if the default value is returned it means that the
flag for that key is not defined. This then means that we can see the
flag name in the dev console for that request. Without doing this, it's
impossible to see what the unknown flag key is in the dev console.

This is a bit of a workaround until LD improve their dev console.