Add new acceptance tests to better cover functionality
The two tests exercising the JS `variation` helper are commented out
because the helper does not work. However, it is going to be re-written
with a new approach in a follow up PR so this is ok for now. By the time
we release a new version of the addon this will work.