Remove babel transform plugin by default
The babel transform that replaces the `variation` JS helper is
problematic. It mostly works, however, detecting whether the helper is
used inside a computed property is buggy and I'm yet to get it to a
point that I feel is solid to be used by default in this addon.

Also, this helper is essentially syntactic sugar and does feel like
something that should be an "opt in".

Therefore, this commit removes the babel transform from being added by
default and, instead, gives users instructions on how they can include
it in their app if they wish to use it.

This is a breaking change.